Move /auth/lichess -> /signin/lichess
This commit also moves the the callback login from
auth.route.js to user.controller.js
