+++ server/src/app.js
@@ -29,7 +29,6 @@ app.use(passport.initialize());
 app.use('/api', require('./routes/user.route'));
 app.use('/api', require('./routes/profile.route'));
 app.use('/api', require('./routes/game.route'));
-app.use('/api/auth', require('./routes/auth.route'));
 
 // Running the server at port PORT or default 8000
 const port = process.env.PORT || 8000;
